[#310] Apply SRP to various git commands (#408)
All the git commands are embedded in the CommandRunner class.

This violates the Single Responsibility Principle as CommandRunner
should only be responsible for the execution, and not the type of
commands.

As a step towards deprecating the use of CommandRunner, let's export
the commands to respective classes and remove the direct dependency on
CommandRunner.
