+++ src/main/java/reposense/authorship/FileInfoAnalyzer.java
@@ -17,9 +17,9 @@ import reposense.authorship.analyzer.CheckStyleParser;
 import reposense.authorship.model.FileInfo;
 import reposense.authorship.model.FileResult;
 import reposense.authorship.model.LineInfo;
+import reposense.git.GitBlame;
 import reposense.model.Author;
 import reposense.model.RepoConfiguration;
-import reposense.system.CommandRunner;
 import reposense.system.LogsManager;
 
 /**
@@ -100,7 +100,7 @@ public class FileInfoAnalyzer {
      * Returns the analysis result from running git blame on {@code filePath}.
      */
     private static String getGitBlameResult(RepoConfiguration config, String filePath) {
-        return CommandRunner.blameRaw(config.getRepoRoot(), filePath);
+        return GitBlame.blame(config.getRepoRoot(), filePath);
     }
 
     /**
