+++ src/main/java/reposense/git/GitCloneException.java
@@ -1,0 +1,8 @@
+package reposense.git;
+
+
+public class GitCloneException extends Exception {
+    public GitCloneException(Exception e) {
+        super(e.getMessage());
+    }
+}
