--- src/main/java/reposense/git/GitDownloaderException.java
@@ -1,8 +1,0 @@
-package reposense.git;
-
-
-public class GitDownloaderException extends Exception {
-    public GitDownloaderException(Exception e) {
-        super(e.getMessage());
-    }
-}
