+++ src/main/java/reposense/git/GitShortlog.java
@@ -1,24 +1,29 @@
 package reposense.git;
 
+import static reposense.system.CommandRunner.runCommand;
+
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
 
 import reposense.model.Author;
 import reposense.model.RepoConfiguration;
-import reposense.system.CommandRunner;
 
 /**
- * Executes and extracts information from git-shortlog related commands.
+ * Contains git shortlog related functionalities.
+ * Git shortlog provides a summary of git log output.
  */
 public class GitShortlog {
 
     /**
      * Extracts all the author identities from the repository and date range given in {@code config}.
      */
-    public static List<Author> extractAuthorsFromLog(RepoConfiguration config) {
-        String summary = CommandRunner.getShortlogSummary(
+    public static List<Author> getAuthors(RepoConfiguration config) {
+        String summary = getShortlogSummary(
                 config.getRepoRoot(), config.getSinceDate(), config.getUntilDate());
 
         if (summary.isEmpty()) {
@@ -29,5 +34,14 @@ public class GitShortlog {
         return Arrays.stream(lines)
                 .map(line -> new Author(line.split("\t")[1]))
                 .collect(Collectors.toList());
+    }
+
+    private static String getShortlogSummary(String root, Date sinceDate, Date untilDate) {
+        Path rootPath = Paths.get(root);
+        String command = "git log --pretty=short";
+        command += GitUtil.convertToGitDateRangeArgs(sinceDate, untilDate);
+        command += " | git shortlog --summary";
+
+        return runCommand(rootPath, command);
     }
 }
