+++ src/test/java/reposense/git/GitShortlogTest.java
@@ -1,7 +1,5 @@
 package reposense.git;
 
-import static reposense.git.GitShortlog.extractAuthorsFromLog;
-
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.List;
@@ -16,39 +14,39 @@ import reposense.util.TestUtil;
 public class GitShortlogTest extends GitTestTemplate {
 
     @Test
-    public void extractAuthorsFromLog_validRepoNoDateRange_success() {
+    public void getAuthors_validRepoNoDateRange_success() {
         List<Author> expectedAuthorList = new ArrayList<>();
         expectedAuthorList.add(new Author("Eugene Peh"));
         expectedAuthorList.add(new Author("eugenepeh"));
         expectedAuthorList.add(new Author("fakeAuthor"));
         expectedAuthorList.add(new Author("harryggg"));
 
-        List<Author> actualAuthorList = extractAuthorsFromLog(config);
+        List<Author> actualAuthorList = GitShortlog.getAuthors(config);
 
         Assert.assertEquals(expectedAuthorList.size(), actualAuthorList.size());
         Assert.assertEquals(expectedAuthorList, actualAuthorList);
     }
 
     @Test
-    public void extractAuthorsFromLog_validRepoDateRange_success() {
+    public void getAuthors_validRepoDateRange_success() {
         List<Author> expectedAuthorList = new ArrayList<>();
 
         expectedAuthorList.add(new Author("eugenepeh"));
         config.setSinceDate(TestUtil.getDate(2018, Calendar.MAY, 5));
         config.setUntilDate(TestUtil.getDate(2018, Calendar.MAY, 10));
 
-        List<Author> actualAuthorList = extractAuthorsFromLog(config);
+        List<Author> actualAuthorList = GitShortlog.getAuthors(config);
 
         Assert.assertEquals(expectedAuthorList.size(), actualAuthorList.size());
         Assert.assertEquals(expectedAuthorList, actualAuthorList);
     }
 
     @Test
-    public void extractAuthorsFromLog_validRepoDateOutOfRange_success() {
+    public void getAuthors_validRepoDateOutOfRange_success() {
         config.setSinceDate(TestUtil.getDate(2018, Calendar.JUNE, 1));
         config.setUntilDate(TestUtil.getDate(2018, Calendar.JUNE, 20));
 
-        List<Author> actualAuthorList = extractAuthorsFromLog(config);
+        List<Author> actualAuthorList = GitShortlog.getAuthors(config);
 
         Assert.assertTrue(actualAuthorList.isEmpty());
     }
